+++ EMDA/plotters.py
@@ -1,5 +1,10 @@
 import matplotlib.pyplot as plt
 from numpy import absolute as abs
+from numpy import average, std, arange, array
+import numpy as np
+
+from .exceptions import NotCompatibleAnalysisForPlotterError, NotCompatibleMeasureForPlotterError
+from .tools import get_dictionary_structure
 
 """
 TO BUILD:
@@ -7,70 +12,494 @@ TO BUILD:
     - [] 
 """
 
-
-def plot_values(self, measure_name, out_name=None):
+def plot_measure(self, measure_name, same_y : bool = True, same_x : bool = True, axis_label_everywhere : bool =False, combine_replicas : bool =False, width_per_replica : float = 4, height_per_variant : float = 4, out_name=False):
     """
     DESCRIPTION:
-        P
+        Function for plotting all measures in a Measure class. Only suitable for frame-wise lists 
+        like the obtained with distance, angle, dihedral, planar angle and RMSD.
 
+    OPTIONS:
+        - measure_name:             [str]        name of the measure listed as EMDA.measures key
+        - same_y, same_x:           [bool]       shares the y and/or x among all plots in the same row/column, so plots have the same axis dimensions
+        - axis_label_everywhere:    [bool]       Adds the x and y axis labels to all the subplots instead of only to the ones at the left and bottom
+        - out_name:                 [pseudobool] False by default. If a string is given, the plot will be saved.
+        - 
     """
 
-    plt.plot(self.measure[measure_name].result)
+    y_labels = {
+        "distance" : "Distance (Å)",
+        "RMSD" : "RMSD (Å)",
+        "angle" : "Angle (°)",
+        "planar_angle" : "Planar angle (°)",
+        "dihedral" : "Dihedral angle (°)",
+        "contacts_amount": "Number of contacts"
+    }
 
-    plt.show()
-
-    if out_name != None:
-        pass
-
-
-def plot_contacts_frequency(self, analysis_name, out_name=None, fill_empty=False, width_plot=0.5):
-    """
-    DESCRIPTION:
-        Plotter that takes the result of a contacts_frequency analysis and plots each interaction as a bar plot
-    """
-
-    if self.analyses[analysis_name].type != "contacts_frequency":
-        #raise 
-        pass
-
-    if fill_empty:
-        all_protein = self.universe.select_atoms('protein')
-        for resname, resid in zip(list(all_protein.residues.resnames), list(all_protein.residues.resids)):
-            if f"{str(resname)}{str(resid)}" not in list(self.analyses[analysis_name].result.keys()):
-                self.analyses[analysis_name].result[f"{str(resname)}{str(resid)}"] = 0
-
-        plt.figure(figsize=(len(self.analyses[analysis_name].result) * width_plot, 5))
-
-        plt.bar(
-            [int(res[3:]) for res in list(self.analyses[analysis_name].result.keys())],
-            #list(self.analyses[analysis_name].result.keys()),
-            [self.analyses[analysis_name].result[res] for res in list(self.analyses[analysis_name].result.keys())]
-            #list(self.analyses[analysis_name].result.values())
-        )
-    
-        plt.xticks(
-        [int(res[3:]) for res in list(self.analyses[analysis_name].result.keys()) if self.analyses[analysis_name].result[res] != 0],
-        [res for res in list(self.analyses[analysis_name].result.keys()) if self.analyses[analysis_name].result[res] != 0],
-        rotation=90
-        )
-
-    elif not fill_empty:
-        plt.figure(figsize=(len(self.analyses[analysis_name].result) * width_plot, 5))
-
-        plt.bar(
-            list(self.analyses[analysis_name].result.keys()),
-            list(self.analyses[analysis_name].result.values())
-        )
-
-
-        plt.xticks(rotation=45, ha="right")
+    # Check if plotting as plotter or as class' method
+    if measure_name == None:
+        measure_obj = self
+    else :
+        measure_obj = self.measures[measure_name]
 
     
+    if measure_obj.type not in ("distance", "angle", "dihedral", "RMSD", "planar_angle", "contacts_amount"):
+        raise NotCompatibleMeasureForPlotterError
+    if measure_obj.type == "contacts_amounts" and measure_obj.options["mode"] not in ("contacts"):
+        raise NotCompatibleMeasureForPlotterError
+    
+    variants = len(measure_obj.result)
+
+    if combine_replicas:
+        max_replicas = 1
+    else :
+        max_replicas = max([ len(measure_obj.result[variant]) for variant in list(measure_obj.result) ])
+
+    #fig, axs = plt.subplots(ncols=variants, nrows=max_replicas, sharey=same_y, sharex=same_x) --> axs[r_num, v_num]
+    # plotting replicas in X axis and variant in Y axis
+    fig, axs = plt.subplots(
+        ncols=max_replicas, nrows=variants, 
+        sharey=same_y, sharex=same_x, 
+        figsize=(max_replicas*width_per_replica, variants*height_per_variant)
+    )
+
+    # Check if only one variant
+
+    if variants == 1 and max_replicas == 1:
+        variant = list(measure_obj.result.keys())[0]
+        replica = list(measure_obj.result[variant].keys())[0]
+
+        axs.plot(
+            range(1, len(measure_obj.result[variant][replica])+1),
+            measure_obj.result[variant][replica],
+            c = f"C0",
+            label=replica
+            )
+        
+        axs.set_ylabel(y_labels[measure_obj.type])
+        axs.set_xlabel("Frame")
+    
+    else :
+        # If there's only one replica, treat it as merged
+        if max_replicas == 1:
+            combine_replicas = True
+
+        for v_num, variant in enumerate(list(measure_obj.result.keys())):
+            for r_num, replica in enumerate(list(measure_obj.result[variant].keys())):
+
+                if combine_replicas:
+                    axs[v_num].plot(
+                    range(1, len(measure_obj.result[variant][replica])+1),
+                    measure_obj.result[variant][replica],
+                    c = f"C{v_num}",
+                    label=replica
+                    )
+
+                    if r_num == 0 or axis_label_everywhere:
+                        axs[v_num].set_ylabel(y_labels[measure_obj.type])
+
+                    if v_num == variants-1 or axis_label_everywhere:
+                        axs[v_num].set_xlabel("Frame")
+                        
+                    axs[v_num].set_title(f"{variant}, replicas {', '.join(list(measure_obj.result[variant].keys()))}")  
+
+                elif variants == 1:
+                    axs[r_num].plot(
+                    range(1, len(measure_obj.result[variant][replica])+1),
+                    measure_obj.result[variant][replica],
+                    c = f"C{v_num}",
+                    label=replica
+                    )
+
+                    if r_num == 0 or axis_label_everywhere:
+                        axs[r_num].set_ylabel(y_labels[measure_obj.type])
+
+                    if v_num == variants-1 or axis_label_everywhere:
+                        axs[r_num].set_xlabel("Frame")
+                        
+                    axs[r_num].set_title(f"{variant}, {replica}")  
+
+                else :
+                    axs[v_num, r_num].plot(
+                        range(1, len(measure_obj.result[variant][replica])+1),
+                        measure_obj.result[variant][replica],
+                        c = f"C{v_num}"
+                        )
+                    
+                    if r_num == 0 or axis_label_everywhere:
+                        axs[v_num, r_num].set_ylabel(y_labels[measure_obj.type])
+                    
+                    if v_num == variants-1 or axis_label_everywhere:
+                        axs[v_num, r_num].set_xlabel("Frame")
+                    
+                    axs[v_num, r_num].set_title(f"{variant}, {replica}")
+
+            if r_num != max_replicas-1:
+                for r_num_ in range(r_num, max_replicas):
+                    if v_num == variants-1 or axis_label_everywhere:
+                        axs[v_num, r_num_].plot()
+                        axs[v_num, r_num_].set_xlabel("Frame")
+
+            
+
+
+    if measure_name == None:
+        fig.suptitle("Plots for " + r"$\bf{%s}$" % self.name.replace('_', '\_') +  " Measure")
+
+    else :
+        fig.suptitle("Plots for " + r"$\bf{%s}$" % measure_name.replace('_', '\_') +  " Measure")
+
+    
+    fig.tight_layout()
+
+    #if merge_replica_plots:
+    #    plt.legend()
+
+    if out_name != False and isinstance(out_name, str):
+        if not out_name.endswith((".png", ".jpg", ".jpeg", ".tiff")):
+            out_name = ".".join(out_name.split('.')[:-1]) + ".png"
+
+        plt.savefig(out_name, dpi=300, bbox_inches='tight')
 
     plt.show()
     plt.close()
 
 
+
+def plot_NACs(self, analysis_name, merge_replicas=False, percentage=False, error_bar=True, bar_width=0.1, width=None, title=None, out_name=False):
+    """
+    DESCRIPTION:
+        Function for plotting NACs (or value-type Analysis) as a bar plot where all the variants are compared
+    """
+
+    # Check if plotting as plotter or as class' method
+    if analysis_name == None:
+        analysis_obj = self
+    else :
+        analysis_obj = self.analyses[analysis_name]
+
+    if analysis_obj.type not in ('value', 'NACs'):
+        raise NotCompatibleMeasureForPlotterError
+
+    # create pyplot subplot
+    fig, ax = plt.subplots(ncols=1,nrows=1)
+    
+    # Check width
+    if width == None:
+        fig.set_figwidth(10*bar_width*len(analysis_obj.result))
+    elif isinstance(width, (float, int)):
+        fig.set_figwidth(width)
+    elif width.lower() in ('auto', 'automatic'):
+        pass
+    else :
+        print("width value ({width}) cannot be understood, using default value.")
+
+    # run code if merging replicas
+    if merge_replicas:
+        max_replicas = max([ len(analysis_obj.result[variant]) for variant in list(analysis_obj.result) ])
+        for v_num, variant in enumerate(list(analysis_obj.result.keys())):
+            # calculate the avg value for the bar (it height)
+            if percentage:
+                avgs = average([analysis_obj.result[variant][replica].count(True)*100/len(analysis_obj.result[variant][replica]) for replica in list(analysis_obj.result[variant].keys())])
+            elif not percentage:
+                avgs = average([analysis_obj.result[variant][replica].count(True) for replica in list(analysis_obj.result[variant].keys())])
+
+            # plot a bar for each replica
+            ax.bar(variant, avgs, bar_width*max_replicas, color = f"C{v_num}")
+            if error_bar and max_replicas != 1:
+                ax.errorbar(variant, avgs, 
+                        yerr=std([analysis_obj.result[variant][replica].count(True) for replica in list(analysis_obj.result[variant].keys())]), 
+                        color = f"k",
+                        solid_capstyle='butt',
+                        capsize=bar_width*72 # in to pt is 72, 5 is to make it wider
+                        )
+
+    elif not merge_replicas: 
+        # Step 1: Process data
+        # Counting True values for each replica in each variant
+        true_counts = {variant: [sum(replica) for replica in replicas.values()] for variant, replicas in analysis_obj.result.items()}
+        all_counts  = {variant: [len(replica) for replica in replicas.values()] for variant, replicas in analysis_obj.result.items()}
+
+        # Find the maximum number of replicas to standardize the data structure
+        max_replicas = max(len(replicas) for replicas in true_counts.values())
+
+        # Initialize a list to hold the count of Trues for each replica in each variant
+        counts_by_replica = {f'R{i+1}': [] for i in range(max_replicas)}
+
+        # Populate the counts by iterating through each variant and replica
+        v_num = 0
+        for variant in true_counts.values():
+            for i, count in enumerate(variant):
+                if percentage:
+                    counts_by_replica[f'R{i+1}'].append(count*100/all_counts[list(true_counts.keys())[v_num]][i])
+                else :
+                    counts_by_replica[f'R{i+1}'].append(count)
+            v_num += 1
+            # If a variant has fewer replicas, append 0 to ensure equal length lists
+            for i in range(len(variant), max_replicas):
+                counts_by_replica[f'R{i+1}'].append(0)
+
+        # Step 2: Organize data for plotting
+        labels = list(true_counts.keys())  # Variant names
+        num_variants = len(true_counts)
+        x = arange(num_variants)  # the label locations
+    
+
+        for i, (replica, counts) in enumerate(counts_by_replica.items()):
+            ax.bar(x + i*bar_width, counts, bar_width, label=replica)
+
+        # Add some text for labels, title, and custom x-axis tick labels, etc.
+            
+        ax.set_xticks(x + bar_width * (max_replicas / 2 - 0.5))
+        ax.set_xticklabels(labels)
+       
+        ax.legend()
+
+
+    if percentage:
+        ax.set_ylim([0,100])
+
+    
+    if title == None:
+        ax.set_title('Plot for ' + r"$\bf{%s}$" % analysis_obj.name.replace('_', '\_') + " Analysis")
+    elif title == '':
+        pass
+    else :
+        ax.set_title(title)
+
+    if percentage:
+        ax.set_ylabel('Percentage of NACs (%)')
+        ax.set_yticks(range(0,101, 10))
+
+    elif not percentage:
+        ax.set_ylabel('Number of NACs')
+    
+#    if not merge_replicas:
+#        ax.set_xticks(x + bar_width * (max_replicas / 2 - 0.5))
+#        ax.set_xticklabels(labels)
+        
+    ax.grid(axis='x')
+    ax.set_xlabel('Variant')
+        
+    if out_name != False and isinstance(out_name, str):
+        if not out_name.endswith((".png", ".jpg", ".jpeg", ".tiff")):
+            out_name = ".".join(out_name.split('.')[:-1]) + ".png"
+
+        plt.savefig(out_name, dpi=300, bbox_inches='tight')
+
+    plt.show()
+    plt.close()
+            
+
+
+def plot_contacts_frequency(
+        #self, analysis_name, fill_empty=False, width_plot=0.5, out_name=None
+        self, analysis_name, same_y : bool = True, same_x : bool = True, axis_label_everywhere : bool =False, merge_replicas : bool =False, error_bar=True, bar_width=0.8, errorbar_width=5 , width_per_replica : float = 4, height_per_variant : float = 4, residue_label_rotation=45, out_name=False
+        ):
+    """
+    DESCRIPTION:
+        Plotter that takes the result of a contacts_frequency analysis and plots each interaction as a bar plot
+
+    TODO:
+        - [ ] Add normalisation to max per replica
+    """
+
+
+    #y_labels = {
+        #"distance" : "Distance (Å)",
+        #"RMSD" : "RMSD (Å)",
+        #"angle" : "Angle (°)",
+        #"planar_angle" : "Planar angle (°)",
+        #"dihedral" : "Dihedral angle (°)"
+    #}
+
+    # Check if plotting as plotter or as class' method
+    if analysis_name == None:
+        analysis_obj = self
+    else :
+        analysis_obj = self.analyses[analysis_name]
+
+    if analysis_obj.type not in ("contacts_frequency") and analysis_obj.options['mode'] not in ('contacts'):
+        raise NotCompatibleAnalysisForPlotterError
+
+    variants = len(analysis_obj.result)
+    max_replicas = max([ len(analysis_obj.result[variant]) for variant in list(analysis_obj.result) ])
+
+    if merge_replicas:
+        # if only one replica, no merge is possible.
+        if max_replicas == 1:
+            merge_replicas = False
+
+        else:
+            max_replicas = 1
+            merged_replicas = {}
+            for variant in list(analysis_obj.result.keys()):
+                merged_replicas[variant] = {}
+                for replica in list(analysis_obj.result[variant].keys()):
+                    for residue in list(analysis_obj.result[variant][replica].keys()):
+                        merged_replicas[variant][residue] = []
+
+            for variant in list(merged_replicas.keys()):
+                for residue in list(merged_replicas[variant].keys()):
+                    for replica in list(analysis_obj.result[variant].keys()):
+                        if residue not in list(analysis_obj.result[variant][replica].keys()):
+                            merged_replicas[variant][residue].append(0)
+                        else :
+                            merged_replicas[variant][residue].append(analysis_obj.result[variant][replica][residue])
+
+            avgs, errs = {}, {}
+            for variant in list(merged_replicas.keys()):
+                avgs[variant], errs[variant] = {}, {}
+
+                for residue in merged_replicas[variant]:
+                    avgs[variant][residue] = average(merged_replicas[variant][residue])
+                    errs[variant][residue] =     std(merged_replicas[variant][residue])
+
+    #fig, axs = plt.subplots(ncols=variants, nrows=max_replicas, sharey=same_y, sharex=same_x) --> axs[r_num, v_num]
+    # plotting replicas in X axis and variant in Y axis
+    fig, axs = plt.subplots(
+        ncols=max_replicas, nrows=variants, 
+        sharey=same_y, sharex=same_x, 
+        figsize=(max_replicas*width_per_replica, variants*height_per_variant)
+    )
+
+    # Check if only one variant
+    if variants == 1 and max_replicas == 1:
+        variant = list(analysis_obj.result.keys())[0]
+        replica = list(analysis_obj.result[variant].keys())[0]
+
+        axs.bar(
+            list(analysis_obj.result[variant][replica].keys()),
+            list(analysis_obj.result[variant][replica].values()),
+            color = f"C0",
+            #label=replica
+            )
+        
+        #axs.set_ylabel(y_labels[analysis_obj.type])
+        axs.set_xlabel("Residue")
+        #axs.set_xticks(rotation=45)
+    
+    else :
+
+        for v_num, variant in enumerate(list(analysis_obj.result.keys())):
+            for r_num, replica in enumerate(list(analysis_obj.result[variant].keys())):
+
+                if merge_replicas:
+                    axs[v_num].bar(
+                        list(avgs[variant].keys()),
+                        list(avgs[variant].values()),
+                        bar_width,
+                        color = f"C{v_num}",
+                    )
+
+                    if error_bar:
+                        #if bar_width < 1:
+                        #    capsize = bar_width*15
+                        #elif bar_width >= 1:
+                        #    capsize = bar_width*2
+                        
+                        axs[v_num].errorbar(
+                            list(avgs[variant].keys()),
+                            list(avgs[variant].values()),
+                            yerr=list(errs[variant].values()),
+                            xerr=None,
+                            ls='none',
+                            color = f"k",
+                            solid_capstyle='butt',
+                            capsize=errorbar_width #capsize #bar_width*5 #72/2 # in to pt is 72, 5 is to make it wider
+                        )
+
+                    #if r_num == 0 or axis_label_everywhere:
+                    #    axs[v_num].set_ylabel(y_labels[analysis_obj.type])
+
+                    if v_num == variants-1 or axis_label_everywhere:
+                        axs[v_num].set_xlabel("Residue")
+                        
+                    #axs[v_num].set_xticks(rotation=45)
+                    axs[v_num].set_title(f"{variant}, average of replicas {', '.join(list(analysis_obj.result[variant].keys()))}")  
+
+                elif max_replicas == 1:
+                    axs[v_num].bar(
+                    list(analysis_obj.result[variant][replica].keys()),
+                    list(analysis_obj.result[variant][replica].values()),
+                    bar_width,
+                    color = f"C{v_num}",
+                    label=replica
+                    )
+
+                    #if r_num == 0 or axis_label_everywhere:
+                    #    axs[v_num].set_ylabel(y_labels[analysis_obj.type])
+
+                    if v_num == variants-1 or axis_label_everywhere:
+                        axs[v_num].set_xlabel("Residue")
+                        
+                    #axs[v_num].set_xticks(rotation=45)
+                    axs[v_num].set_title(f"{variant}, replicas {', '.join(list(analysis_obj.result[variant].keys()))}")  
+
+                elif variants == 1:
+                    axs[r_num].bar(
+                        list(analysis_obj.result[variant][replica].keys()),
+                        list(analysis_obj.result[variant][replica].values()),
+                        bar_width,
+                        color = f"C{v_num}",
+                        #label=replica
+                    )
+
+                    #if r_num == 0 or axis_label_everywhere:
+                    #    axs[r_num].set_ylabel(y_labels[analysis_obj.type])
+
+                    if v_num == variants-1 or axis_label_everywhere:
+                        axs[r_num].set_xlabel("Residue")
+                        
+                    axs[r_num].set_title(f"{variant}, {replica}")  
+
+                else :
+                    axs[v_num, r_num].bar(
+                        list(analysis_obj.result[variant][replica].keys()),
+                        list(analysis_obj.result[variant][replica].values()),
+                        bar_width,
+                        color = f"C{v_num}"
+                        )
+                    
+                    #if r_num == 0 or axis_label_everywhere:
+                    #    axs[v_num, r_num].set_ylabel(y_labels[analysis_obj.type])
+                    
+                    if v_num == variants-1 or axis_label_everywhere:
+                        axs[v_num, r_num].set_xlabel("Residue")
+                    
+                    axs[v_num, r_num].set_title(f"{variant}, {replica}")
+
+            if r_num != max_replicas-1:
+                for r_num_ in range(r_num, max_replicas):
+                    if v_num == variants-1 or axis_label_everywhere:
+                        axs[v_num, r_num_].plot()
+                        axs[v_num, r_num_].set_xlabel("Residue")
+
+            
+    fig.autofmt_xdate(rotation=residue_label_rotation)
+
+
+    if analysis_name == None:
+        fig.suptitle("Plots for " + r"$\bf{%s}$" % self.name.replace('_', '\_') +  " Measure")
+
+    else :
+        fig.suptitle("Plots for " + r"$\bf{%s}$" % analysis_name.replace('_', '\_') +  " Measure")
+
+    if analysis_obj.options['percentage']:
+        plt.setp(axs, ylim=(0,100))
+    
+    fig.tight_layout()
+
+    #if merge_replica_plots:
+    #    plt.legend()
+
+    if out_name != False and isinstance(out_name, str):
+        if not out_name.endswith((".png", ".jpg", ".jpeg", ".tiff")):
+            out_name = ".".join(out_name.split('.')[:-1]) + ".png"
+
+        plt.savefig(out_name, dpi=300, bbox_inches='tight')
+
+    plt.show()
+    plt.close()
 
 
 def ext_plot_contacts_frequencies_differences(
@@ -96,7 +525,7 @@ def ext_plot_contacts_frequencies_differences(
         - return_results:       returns the dictionary with the compared results.
         - return_labels:        returns the list of labels of the obtained results.
         - width_plot:           sets the width per bar of the output plot.
-        - save_plot:            pseudoboolean value to save the generated plot. If True, it will be stored as 'contacts_frequencies_diffs.png'. 
+        - save_plot:            pseudoboolean value to save the generated plot. If True, it will be stored as 'contacts_frequencies_diffs.png'.
                                 If a string is given, it will be used as the file name's with the .png extension.
 
     BUILDING NOTES:
@@ -180,11 +609,13 @@ def ext_plot_contacts_frequencies_differences(
             elif int(k_[0]) == int(k_[1]) - 1:
                 del important_contacts[k]
 
-
     # Remove redundant
     for k in list(important_contacts.keys()):
         for k_ in list(important_contacts.keys()):
-            if k.split('-')[0] == k_.split('-')[1] and k.split('-')[1] == k_.split('-')[0]:
+            if (
+                k.split("-")[0] == k_.split("-")[1]
+                and k.split("-")[1] == k_.split("-")[0]
+            ):
                 del important_contacts[k]
 
     if len(important_contacts) == 0:
@@ -222,16 +653,13 @@ def ext_plot_contacts_frequencies_differences(
     else:
         plt.ylabel("Frequency (number of)")
 
-    plt.show()
-
     if save_plot:
-        plt.savefig('contacts_frequencies_diffs.png', dpi=300, bbox_inches='tight')
+        plt.savefig("contacts_frequencies_diffs.png", dpi=300, bbox_inches="tight")
 
     elif isinstance(save_plot, str):
-        plt.savefig(save_plot + '.png', dpi=300, bbox_inches='tight')
+        plt.savefig(save_plot + ".png", dpi=300, bbox_inches="tight")
 
-        
-
+    plt.show()
     plt.close()
 
     if return_results and return_labels:
@@ -240,4 +668,3 @@ def ext_plot_contacts_frequencies_differences(
         return important_contacts
     elif return_labels:
         return return_labels
-